data_augmentation_research/requirements.txt
streamlit==1.2.0
streamlit-autorefresh==0.0.1
numpy==1.21.4
opencv-python-headless==4.5.4.60
pandas==1.3.4
Pyrebase4==4.5.0
